Rename *.js to *.ts and opt out of type checking using any for now
