+++ src/assigners/MonthFormatTokenAssigner.ts
@@ -1,15 +1,19 @@
-const ISODayOfWeekFormatTokenAssigner = (function() {
-	const Assigner = {};
+const MonthFormatTokenAssigner = (function() {
+	const Assigner: any = {};
 
 	// Assigner name
-	Assigner.name = 'ISODayOfWeekFormatTokenAssigner';
+	Assigner.name = 'MonthFormatTokenAssigner';
 
 	// Assigner type
-	Assigner.type= 'isoDayOfWeek';
+	Assigner.type= 'month';
 
 	// Regexp for matching the format token 
 	Assigner.map = new Map();
-	Assigner.map.set(/[1-7]/, 'E');
+	Assigner.map.set(/\d{1,2}/, 'M');
+	Assigner.map.set(/\d{2}/, 'MM');
+	Assigner.map.set(/\d{1,2}(?:st|nd|rd|th)/, 'Mo');
+	Assigner.map.set(/^(Jan|Feb|Mar|Apr|May|Jun|Jul|Aug|Sep|Oct|Nov|Dec)$/, 'MMM');
+	Assigner.map.set(/^(January|February|March|April|May|June|July|August|September|October|November|December)$/, 'MMMM');
 
 	/**
 	 * Tests whether token type is same as
@@ -38,6 +42,7 @@ const ISODayOfWeekFormatTokenAssigner = (function() {
 	};
 
 	return Assigner;
-})(); 
+})();
 
-export default ISODayOfWeekFormatTokenAssigner; 
+export default MonthFormatTokenAssigner; 
+
