+++ src/parsers/Token.ts
@@ -1,8 +1,14 @@
 export default class Token {
+
+	value: any;
+	type: any;
+	index: any;
+	format: any;
+
 	/**
 	 * Constructor.
 	 */
-	constructor({ value, type, index, format }) {
+	constructor({ value, type, index, format }: any) {
 		this.value = value;
 		this.type = type;
 		this.format = format
